Agregando controlador para trueque.

+++ src/main/java/com/ingseoft/swapp/Repository/TruequeRepository.java
@@ -7,5 +7,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.ingseoft.swapp.Model.Trueque;
 
 public interface TruequeRepository extends JpaRepository<Trueque,Integer> {
-    List<Trueque> findByIdTrueque(Long IdTrueque);
+    List<Trueque> findByIdTrueque(Integer IdTrueque);
 }
